backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Scheduler
apscheduler==3.10.4

# HTTP requests
httpx==0.27.2
requests==2.32.3

# Data processing
pandas
numpy
scipy

# Supabase
supabase==2.9.0

# Technical analysis
pandas-ta



# Google Gemini (Autoresearch)
google-generativeai

# EODHD (Seed data + EURIBOR)
eodhd==1.0.26

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
loguru==0.7.2
